--- paimon-flink/paimon-flink-1.14/src/main/java/org/apache/paimon/flink/source/StaticFileStoreSplitEnumerator.java
@@ -1,46 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.paimon.flink.source;
-
-import org.apache.paimon.Snapshot;
-import org.apache.paimon.flink.source.assigners.SplitAssigner;
-
-import org.apache.flink.api.connector.source.SplitEnumerator;
-import org.apache.flink.api.connector.source.SplitEnumeratorContext;
-
-import javax.annotation.Nullable;
-
-/** A {@link SplitEnumerator} implementation for {@link StaticFileStoreSource} input. */
-public class StaticFileStoreSplitEnumerator extends StaticFileStoreSplitEnumeratorBase {
-
-    public StaticFileStoreSplitEnumerator(
-            SplitEnumeratorContext<FileStoreSourceSplit> context,
-            @Nullable Snapshot snapshot,
-            SplitAssigner splitAssigner) {
-        this(context, snapshot, splitAssigner, null);
-    }
-
-    public StaticFileStoreSplitEnumerator(
-            SplitEnumeratorContext<FileStoreSourceSplit> context,
-            @Nullable Snapshot snapshot,
-            SplitAssigner splitAssigner,
-            @Nullable DynamicPartitionFilteringInfo dynamicPartitionFilteringInfo) {
-        super(context, snapshot, splitAssigner);
-    }
-}
