+++ paimon-flink/paimon-flink-common/src/main/java/org/apache/paimon/flink/source/assigners/DynamicPartitionPruningAssigner.java
@@ -24,7 +24,10 @@ import org.apache.paimon.flink.FlinkRowData;
 import org.apache.paimon.flink.source.FileStoreSourceSplit;
 import org.apache.paimon.table.source.DataSplit;
 
+import org.apache.flink.api.connector.source.SourceEvent;
 import org.apache.flink.table.connector.source.DynamicFilteringData;
+import org.apache.flink.table.connector.source.DynamicFilteringEvent;
+import org.slf4j.Logger;
 
 import javax.annotation.Nullable;
 
@@ -81,6 +84,23 @@ public class DynamicPartitionPruningAssigner implements SplitAssigner {
         return innerAssigner.remainingSplits().stream()
                 .filter(this::filter)
                 .collect(Collectors.toList());
+    }
+
+    public static SplitAssigner createDynamicPartitionPruningAssignerIfNeeded(
+            int subtaskId,
+            SplitAssigner oriAssigner,
+            Projection partitionRowProjection,
+            SourceEvent sourceEvent,
+            Logger logger) {
+        DynamicFilteringData dynamicFilteringData = ((DynamicFilteringEvent) sourceEvent).getData();
+        logger.info(
+                "Received DynamicFilteringEvent: {}, is filtering: {}.",
+                subtaskId,
+                dynamicFilteringData.isFiltering());
+        return dynamicFilteringData.isFiltering()
+                ? new DynamicPartitionPruningAssigner(
+                        oriAssigner, partitionRowProjection, dynamicFilteringData)
+                : oriAssigner;
     }
 
     private boolean filter(FileStoreSourceSplit sourceSplit) {
